Allow uploading of RI to FFE site

+++ ffe/ffe_session.py
@@ -2,6 +2,7 @@ import re
 import time
 import webbrowser
 from datetime import datetime
+from enum import Enum, auto
 from pathlib import Path
 from typing import Any
 
@@ -32,10 +33,17 @@ FEES_EVENT: str = 'ctl00$ContentPlaceHolderMain$CmdFactureHomologation'
 UPLOAD_LINK_ID: str = 'ctl00_ContentPlaceHolderMain_CmdUploadPapi'
 UPLOAD_EVENT: str = UPLOAD_LINK_ID.replace('_', '$')
 UPLOAD_FILE_ID: str = 'ctl00$ContentPlaceHolderMain$UploadPapi'
+UPLOAD_RI_LINK_ID = 'ctl00_ContentPlaceHolderMain_CmdUploadRI'
+UPLOAD_RI_EVENT: str = UPLOAD_RI_LINK_ID.replace('_', '$')
+UPLOAD_RI_FILE_ID = 'ctl00$ContentPlaceHolderMain$UploadRI'
 
 FEES_DIR: Path = Path('fees')
 
-
+class FFEAction(Enum):
+    UPLOAD_PAPI = auto()
+    UPLOAD_PAPI_AND_SET_VISIBLE = auto()
+    UPLOAD_RI = auto()
+    
 class FFESession(Session):
     """A requests session specialized for communication with the FFE website.
     Currently, it relies on hacks, because no API is available."""
@@ -84,7 +92,7 @@ class FFESession(Session):
             content: str = response.content.decode()
             if self.debug:
                 date_str = datetime.strftime(datetime.fromtimestamp(time.time()), '%Y-%m-%d-%H-%M-%S')
-                debug_file = TMP_DIR / f'{url.replace("/", "_")}-{date_str}-raw.html'
+                debug_file = TMP_DIR / f'{url.replace("/", "_").replace(":", "_")}-{date_str}-raw.html'
                 with open(debug_file, 'w') as file:
                     file.write(content)
                 logger.info('Raw content stored to %s', debug_file)
@@ -107,7 +115,7 @@ class FFESession(Session):
         parser.parseStr(html)
         if self.debug:
             date_str = datetime.strftime(datetime.fromtimestamp(time.time()), '%Y-%m-%d-%H-%M-%S')
-            debug_file = TMP_DIR / f'{self.last_url_read.replace("/", "_")}-{date_str}-parsed.html'
+            debug_file = TMP_DIR / f'{self.last_url_read.replace("/", "_").replace(":", "_")}-{date_str}-parsed.html'
             with open(debug_file, 'w', encoding="utf-8") as file:
                 file.write(parser.getHTML())
             logger.info('Parsed content stored to %s', debug_file)
@@ -171,7 +179,7 @@ class FFESession(Session):
         if not self.read_ffe_state(parser, url):
             return False
         self.auth_state: dict[str, str] = {}
-        for id in [SET_VISIBLE_LINK_ID, FEES_LINK_ID, UPLOAD_LINK_ID, ]:
+        for id in [SET_VISIBLE_LINK_ID, FEES_LINK_ID, UPLOAD_LINK_ID, UPLOAD_RI_LINK_ID]:
             tag: AdvancedTag = parser.getElementById(id)
             self.auth_state[id] = tag.innerText if tag else None
             if self.debug:
@@ -247,7 +255,7 @@ class FFESession(Session):
         logger.info('fees OK')
         return
 
-    def upload(self, set_visible: bool):
+    def upload(self, action: FFEAction):
         """Upload the tournament to the FFE admin website."""
         logger.info(
             'Mise à jour du tournoi [%s] (%s) sur le site fédéral :',
@@ -258,43 +266,59 @@ class FFESession(Session):
         if not self._ffe_auth():
             return
         logger.info('auth OK: %s', self.tournament_ffe_url)
+        upload_link_id = UPLOAD_RI_LINK_ID if action == FFEAction.UPLOAD_RI else UPLOAD_LINK_ID
+        upload_event = UPLOAD_RI_EVENT if action == FFEAction.UPLOAD_RI else UPLOAD_EVENT
+        upload_file_id = UPLOAD_RI_FILE_ID if action == FFEAction.UPLOAD_RI else UPLOAD_FILE_ID
         if self.debug:
-            logger.info('> auth_state[%s]=[%s]', UPLOAD_LINK_ID, self.auth_state[UPLOAD_LINK_ID])
-        if self.auth_state[UPLOAD_LINK_ID] is None:
+            logger.info('> auth_state[%s]=[%s]', upload_link_id, self.auth_state[upload_link_id])
+        if self.auth_state[upload_link_id] is None:
             logger.warning('Lien de mise en ligne non trouvé (vérifier que le tournoi n\'est pas terminé)')
             return
         url = FFE_URL + '/MonTournoi.aspx'
         post: dict[str, str] = {
-            '__EVENTTARGET': UPLOAD_EVENT,
+            '__EVENTTARGET': upload_event,
             '__EVENTARGUMENT': '',
             VIEW_STATE_INPUT_ID: self.ffe_state[VIEW_STATE_INPUT_ID],
             VIEW_STATE_GENERATOR_INPUT_ID: self.ffe_state[VIEW_STATE_GENERATOR_INPUT_ID],
             EVENT_VALIDATION_INPUT_ID: self.ffe_state[EVENT_VALIDATION_INPUT_ID],
         }
         date: str = datetime.fromtimestamp(time.time()).strftime("%Y%m%d%H%M%S")
-        tmp_file: Path = TMP_DIR / 'ffe' / f'{self.tournament.file.stem}-{date}{self.tournament.file.suffix}'
-        tmp_file.parents[0].mkdir(parents=True, exist_ok=True)
-        logger.debug('Copie de %s vers %s...', self.tournament.file, tmp_file)
-        tmp_file.write_bytes(self.tournament.file.read_bytes())
-        with PapiDatabase(tmp_file, write=True) as tmp_database:
-            tmp_database: PapiDatabase
-            logger.debug('Suppression des données personnelles des joueur·euses...')
-            tmp_database.delete_players_personal_data()
-            logger.debug('Suppression des forfaits si aucun appariement...')
-            tmp_database.remove_forfeits_if_no_pairings()
-            tmp_database.commit()
-        html: str = self._read_url(url=url, data=post, files={UPLOAD_FILE_ID: tmp_file, })
-        tmp_file.unlink()
+        if action == FFEAction.UPLOAD_RI:
+            if not self.tournament.rules:
+                logger.warning('Aucun fichier spécifié pour le règlement du tournoi [%s].', self.tournament.name)
+                return
+            else:
+                ri_file = Path(self.tournament.rules)
+                logger.info('Chemin du fichier RI: [%s]', ri_file)
+                if not ri_file.exists():
+                    logger.error('Fichier RI non trouvé [%s]', ri_file)
+                else:
+                    html: str = self._read_url(url=url, data=post, files={upload_file_id: ri_file, })
+        else:
+            tmp_file: Path = TMP_DIR / 'ffe' / f'{self.tournament.file.stem}-{date}{self.tournament.file.suffix}'
+            tmp_file.parents[0].mkdir(parents=True, exist_ok=True)
+            logger.debug('Copie de %s vers %s...', self.tournament.file, tmp_file)
+            tmp_file.write_bytes(self.tournament.file.read_bytes())
+            with PapiDatabase(tmp_file, write=True) as tmp_database:
+                tmp_database: PapiDatabase
+                logger.debug('Suppression des données personnelles des joueur·euses...')
+                tmp_database.delete_players_personal_data()
+                logger.debug('Suppression des forfaits si aucun appariement...')
+                tmp_database.remove_forfeits_if_no_pairings()
+                tmp_database.commit()
+            html: str = self._read_url(url=url, data=post, files={upload_file_id: tmp_file, })
+            tmp_file.unlink()
         if not html:
             return
         _, error = self._parse_html_content(html)
         if error:
             return
-        with EventDatabase(self.tournament.event.uniq_id, write=True) as event_database:
-            event_database.set_tournament_last_ffe_upload(self.tournament.id)
-            event_database.commit()
+        if action != FFEAction.UPLOAD_RI:
+            with EventDatabase(self.tournament.event.uniq_id, write=True) as event_database:
+                event_database.set_tournament_last_ffe_upload(self.tournament.id)
+                event_database.commit()
         logger.info('upload OK')
-        if not set_visible:
+        if action != FFEAction.UPLOAD_PAPI_AND_SET_VISIBLE:
             return
         if self.debug:
             logger.info('> auth_state[%s]=[%s]', SET_VISIBLE_LINK_ID, self.auth_state[SET_VISIBLE_LINK_ID])
